Spring Data JPA를 이용한 Entity 설정과 초기 데이터 생성.

+++ inflearn/RESTfulWebServices/restful-web-service/src/main/java/kr/co/bonjin/restfulwebservice/user/UserDaoService.java
@@ -14,9 +14,9 @@ public class UserDaoService {
     private static int usersCount = 0;
 
     static {
-        users.add(new User(1, "Kenneth", new Date(), "pass1", "701010-1111111"));
-        users.add(new User(2, "Alice", new Date(), "pass2", "701010-2222222"));
-        users.add(new User(3, "Elena", new Date(), "pass3", "701010-3333333"));
+        users.add(new User(1L, "Kenneth", new Date(), "pass1", "701010-1111111"));
+        users.add(new User(2L, "Alice", new Date(), "pass2", "701010-2222222"));
+        users.add(new User(3L, "Elena", new Date(), "pass3", "701010-3333333"));
 
         usersCount = users.size();
     }
@@ -36,7 +36,7 @@ public class UserDaoService {
 
     public User save(User user) {
         if (user.getId() == null) {
-            user.setId(++usersCount);
+            user.setId((long) ++usersCount);
         }
         users.add(user);
         return user;
